Update docker file to have a default config location. (#300)
We specify an entrypoint in the Dockerfile but expect the users to
supply the arguments (mainly the --config) parameter. This change adds a
default value for the `--config` parameter so users don't have to
specify it.

Before this change users would have to issue the following command to
run the collector with docker:

    docker run -v $(pwd):/etc/otel -it otel/opentelemetry-collector-contrib:0.3.0 --config /etc/otel/config.yaml

This change reduces it to the following:

    docker run -v $(pwd):/etc/otel -it otel/opentelemetry-collector-contrib:0.3.0

Users can still specify the --config flag to use a different file.

This also makes the Dockerfile a bit more self-documenting.